Renamed sign up middleware to users middleware
